+++ app/src/main/java/com/example/feedme/OrderViewForRestaurants.kt
@@ -20,7 +20,6 @@ class OrderViewForRestaurants : AppCompatActivity() {
         val restaurantId = intent.getStringExtra("RESID").toString()
         val tvRestaurantName = findViewById<TextView>(R.id.TVrestaurantOrdersRestaurantTitle)
         tvRestaurantName.text =  intent.getStringExtra("RESNAME").toString()
-        getOrdersForRestaurant(restaurantId)
         recyclerViewRestaurantOrders = findViewById<RecyclerView>(R.id.RvRestaurantOrders)
         recyclerViewRestaurantOrders.layoutManager = LinearLayoutManager(this)
         val adapter = RestaurantOrdersRVAdapter(this, DataManagerOrders.orders)
@@ -32,31 +31,4 @@ class OrderViewForRestaurants : AppCompatActivity() {
         super.onResume()
         recyclerViewRestaurantOrders.adapter?.notifyDataSetChanged()
     }
-    fun getOrdersForRestaurant(restaurantId: String){
-        DataManagerOrders.orders.clear()
-        var index = 0
-        db.collection("restaurants").document(restaurantId).collection("orders")
-            .get()
-            .addOnSuccessListener { documents ->
-                for (document in documents) {
-                    DataManagerOrders.orders.add(document.toObject())
-                    db.collection("restaurants").document(restaurantId).collection("orders").document(document.id).collection("orderedDishes")
-                        .get()
-                        .addOnSuccessListener { orderItems ->
-                            for(orderItem in orderItems) {
-                                DataManagerOrders.orders.get(index).orderedDishes.add(orderItem.toObject<OrderItem>())
-                                Log.d(ContentValues.TAG, "${orderItem.id} => ${orderItem.data}")
-                            }
-                        }
-                        .addOnFailureListener { exception ->
-                            Log.w(ContentValues.TAG, "Error getting documents: ", exception)
-                        }
-                    Log.d(ContentValues.TAG, "${document.id} => ${document.data}")
-                }
-            }
-            .addOnFailureListener { exception ->
-                Log.w(ContentValues.TAG, "Error getting documents:", exception)
-            }
-    }
-
 }
